backend/simple/contents/requirements.txt
cassandra-driver==3.29.0
grpcio==1.62.0
protobuf==4.25.3
python-dotenv==1.0.1
